chore: rename signet-launch to v1
