+++ app/e2e_include_upgrades.go
@@ -2,12 +2,10 @@
 
 package app
 
-import (
-	"github.com/babylonlabs-io/babylon/app/upgrades/signetlaunch"
-)
+import v1 "github.com/babylonlabs-io/babylon/app/upgrades/v1"
 
 // init is used to include signet upgrade used for e2e testing
 // this file should be removed once the upgrade testing with signet ends.
 func init() {
-	Upgrades = append(Upgrades, signetlaunch.Upgrade)
+	Upgrades = append(Upgrades, v1.Upgrade)
 }
