+++ test/e2e/configurer/config/constants.go
@@ -16,5 +16,5 @@ const (
 	// PropSubmitBlocks estimated number of blocks it takes to submit for a proposal
 	PropSubmitBlocks float32 = 1
 	// Upgrade prop files json
-	UpgradeSignetLaunchFilePath = "/upgrades/signet-launch.json"
+	UpgradeSignetLaunchFilePath = "/upgrades/v1.json"
 )
